+++ src/main/java/com/aerospike/perseus/domain/data/random/SimpleRecordDataProvider.java
@@ -22,10 +22,10 @@ public class SimpleRecordDataProvider implements DataProvider<Record> {
 
     @Override
     public Record next() {
-        var key = random.nextInt();
+        var key = random.nextLong();
         int octet = random.nextInt(16);
-        int keyPlus10 = key+10;
-        int keyPlus20 = key+20;
+        long keyPlus10 = key+10;
+        long keyPlus20 = key+20;
         String country = new Locale("", locales[ random.nextInt( locales.length) ] ).getDisplayCountry();
         long date = between(Instant.now().minus(Period.ofWeeks(300)), Instant.now()).getEpochSecond();
