+++ src/main/java/com/aerospike/perseus/testCases/UpdateTest.java
@@ -9,16 +9,16 @@ import com.aerospike.perseus.utilities.aerospike.AerospikeConfiguration;
 import java.time.Instant;
 import java.util.List;
 
-public class UpdateTest extends Test<Integer>{
+public class UpdateTest extends Test<Long>{
     private final WritePolicy policy;
 
-    public UpdateTest(AerospikeConfiguration conf, KeyProvider<Integer> provider) {
+    public UpdateTest(AerospikeConfiguration conf, KeyProvider<Long> provider) {
         super(conf, provider);
         policy = new WritePolicy();
         policy.recordExistsAction = RecordExistsAction.UPDATE_ONLY;
     }
 
-    protected void execute(Integer key){
+    protected void execute(Long key){
         Bin updated = new Bin("UpdateRes", Instant.now().toString());
         try{
             client.put(policy, getKey(key), updated);
